refactor utility functions, add comments
