+++ habitat-lab/habitat/utils/geometry_utils.py
@@ -115,3 +115,10 @@ def agent_state_target2ref(
     )
 
     return (rotation_in_ref_coordinate, position_in_ref_coordinate)
+
+
+def cosine(v1, v2):
+    return np.clip(np.dot(v1, v2), -1.0, 1.0)
+
+def angle_between(v1, v2):
+    return np.arccos(cosine(v1, v2))
