+++ habitat-lab/habitat/utils/visualizations/utils.py
@@ -226,6 +226,7 @@ def observations_to_image(observation: Dict, info: Dict) -> np.ndarray:
             if obs_k.shape[2] == 1:
                 obs_k = np.concatenate([obs_k for _ in range(3)], axis=2)
             if 'third' not in sensor_name:
+                # Mark the 3x3 square at the center of the frame with red pixels
                 height, width = obs_k.shape[:2]
                 obs_k[height//2 - 3: height//2 + 3, width//2 - 3: width//2 + 3] = [[255, 0, 0]]
             render_obs_images.append(obs_k)
